Add isSandbox flag

+++ TweetNestKit/Session/Session.swift
@@ -56,6 +56,10 @@ public actor Session {
 }
 
 extension Session {
+    static let isSandbox: Bool = {
+        CKContainer.default().value(forKeyPath: "containerID.environment") as? CLongLong == 2
+    }()
+    
     static var containerURL: URL {
         FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: Session.applicationGroupIdentifier)!
     }
@@ -74,29 +78,33 @@ extension Session {
         let containerApplicationSupportURL = containerLibraryURL
             .appendingPathComponent("Application Support")
             .appendingPathComponent(Bundle.tweetNestKit.bundleIdentifier!)
-
-        // Migration START
-        let oldURL = Session.containerURL
-            .appendingPathComponent("Application Support")
-            .appendingPathComponent(Bundle.tweetNestKit.name!)
-        let newURL = containerApplicationSupportURL
-        if FileManager.default.fileExists(atPath: oldURL.path) {
-            do {
-                try FileManager.default.createDirectory(
-                    at: newURL.deletingLastPathComponent(),
-                    withIntermediateDirectories: true,
-                    attributes: nil
-                )
-                try FileManager.default.moveItem(at: oldURL, to: newURL)
-                try FileManager.default.removeItem(at: oldURL.deletingLastPathComponent())
-            } catch {
-                Logger(label: Bundle.tweetNestKit.bundleIdentifier!, category: String(reflecting: Self.self))
-                    .error("\(error as NSError, privacy: .public)")
+        
+        if isSandbox {
+            return containerApplicationSupportURL.appendingPathExtension("sandbox")
+        } else {
+            // Migration START
+            let oldURL = Session.containerURL
+                .appendingPathComponent("Application Support")
+                .appendingPathComponent(Bundle.tweetNestKit.name!)
+            let newURL = containerApplicationSupportURL
+            if FileManager.default.fileExists(atPath: oldURL.path) {
+                do {
+                    try FileManager.default.createDirectory(
+                        at: newURL.deletingLastPathComponent(),
+                        withIntermediateDirectories: true,
+                        attributes: nil
+                    )
+                    try FileManager.default.moveItem(at: oldURL, to: newURL)
+                    try FileManager.default.removeItem(at: oldURL.deletingLastPathComponent())
+                } catch {
+                    Logger(label: Bundle.tweetNestKit.bundleIdentifier!, category: String(reflecting: Self.self))
+                        .error("\(error as NSError, privacy: .public)")
+                }
             }
-        }
-        // Migration END
+            // Migration END
 
-        return containerApplicationSupportURL
+            return containerApplicationSupportURL
+        }
     }
 }
